requirements.txt
psycopg2-binary
pydantic
fastapi
fastapi-camelcase
uvicorn